fix(proto): break early when read error occurs
Because the read methods of ProtocolReader does not return error, the
timing of the error check will occur at the end of a full operation.
This makes it possible to read endlessly when a read error occurs during
methods like parseInfoList().

Break early when reading the conn in a loop to avoid this.

Fixes https://github.com/jfreymuth/pulse/issues/30